Updated error page arrangement
